+++ juniper-recurrence/juniper_recurrence/routers/__init__.py
@@ -9,6 +9,7 @@ from juniper_recurrence.routers.crossval import router as crossval_router
 from juniper_recurrence.routers.dataset import router as dataset_router
 from juniper_recurrence.routers.model import router as model_router
 from juniper_recurrence.routers.predict import router as predict_router
+from juniper_recurrence.routers.snapshots import router as snapshots_router
 from juniper_recurrence.routers.training import router as training_router
 
-__all__ = ["training_router", "predict_router", "model_router", "dataset_router", "crossval_router"]
+__all__ = ["training_router", "predict_router", "model_router", "dataset_router", "crossval_router", "snapshots_router"]